requirements.txt
Flask
ariadne
pymongo
python-dotenv
requests
streamlit
PyJWT
Flask-Bcrypt